(FACT-1910) Remove check that mount point is physical/tmpfs
This removes the check that the returned mount points are physical
devices or temporary filesystems, and instead will return mount points
for all mounts on the system. Removing this check was the original
suggestion for FACT-1519 to allow tmpfs mount points, but instead was
replaced with a flag to allow only tmpfs filesystems. Instead of adding
additional flags for different filesystems to allow, I propose removing
this check entirely. This will expands the data collected by facter and
allow more fine grained system configuration.