Skip over any mount for /proc or /sys
The puppetlabs facter team said that facter should not return the cgroup
and kernel-internal filesystems. This adds a check to skip the mount
point if it is mounted under the proc or sys directories. It keeps the
tmpfs check to prevent breaking any code that looks for the
currently-returned tmpfs mount information under these directories.